Enum vortex_scalar::ScalarValue

pub enum ScalarValue {
    Bool(bool),
    Primitive(PValue),
    Buffer(Buffer),
    BufferString(BufferString),
    List(Arc<[ScalarValue]>),
    Null,
}

Represents the internal data of a scalar value. Must be interpreted by wrapping up with a DType to make a Scalar.

Note that these values can be deserialized from JSON or other formats. So a PValue may not have the correct width for what the DType expects. This means primitive values must be cast on-read.

Variants

Bool(bool)

Primitive(PValue)

Buffer(Buffer)

BufferString(BufferString)

List(Arc<[ScalarValue]>)

Null

Implementations

impl ScalarValue

pub fn is_null(&self) -> bool

pub fn is_instance_of(&self, dtype: &DType) -> bool

pub fn as_bool(&self) -> VortexResult<Option<bool>>

pub fn as_pvalue(&self) -> VortexResult<Option<PValue>>

pub fn as_buffer(&self) -> VortexResult<Option<Buffer>>

pub fn as_buffer_string(&self) -> VortexResult<Option<BufferString>>

pub fn as_list(&self) -> VortexResult<Option<&Arc<[Self]>>>